backend/requirements.txt
requests
mysql-connector-python
python-dotenv
Flask
Flask-Cors
